Beautifully change : adc.c/h

+++ USER/main.c
@@ -16,7 +16,7 @@
 #include "gpio.h"
 #include "wificonf.h"
 #include "queue.h"
-#include "can.h"
+#include "can1.h"
 #include "can2.h"
 #include "timer.h"
 #include "adc.h"
@@ -50,8 +50,10 @@ void Initialization (void)
 	uart_init(115200);
 	/*初始化延时功能*/
 	delay_init(168); 
-//	/* 出厂设置，一般运行不需要*/
-//	setFactory();
+	
+	/* 出厂设置，一般运行不需要*/
+	/*setFactory();*/
+	
 	#if PRINT_UART_LOG
 	printf("System Initing...!\r\n");
 	#endif
@@ -140,13 +142,13 @@ void testAdc(void){
 	AdcTemp = ADC_Read(ADC_MAX_BYTES);
 	#if PRINT_UART_LOG
 //	printf("%d ",(AdcTemp[0]*256+AdcTemp[1]));
-//	printf("%d ",(AdcTemp[1]*256+AdcTemp[2]));
-//	printf("%d ",(AdcTemp[0]*256+AdcTemp[1]));
-//	printf("%d \r\n",(AdcTemp[0]*256+AdcTemp[1]));
-	printf("%.3f   ",(float)(AdcTemp[0]*256+AdcTemp[1])*5.0/32768.0);	
-	printf("%.3f   ",(float)(AdcTemp[2]*256+AdcTemp[3])*5.0/32768.0);	
-	printf("%.3f   ",(float)(AdcTemp[4]*256+AdcTemp[5])*5.0/32768.0);	
-	printf("%.3f   ",(float)(AdcTemp[6]*256+AdcTemp[7])*5.0/32768.0);	
+//	printf("%d ",(AdcTemp[2]*256+AdcTemp[3]));
+//	printf("%d ",(AdcTemp[4]*256+AdcTemp[5]));
+//	printf("%d \r\n",(AdcTemp[6]*256+AdcTemp[7]));
+	printf("%.3f   ",(double)(AdcTemp[0]*256+AdcTemp[1])*5.0/32768.0);	
+	printf("%.3f   ",(double)(AdcTemp[2]*256+AdcTemp[3])*5.0/32768.0);	
+	printf("%.3f   ",(double)(AdcTemp[4]*256+AdcTemp[5])*5.0/32768.0);	
+	printf("%.3f   ",(double)(AdcTemp[6]*256+AdcTemp[7])*5.0/32768.0);	
 	printf("\r\n");
 	delay_ms(1000);
 	#endif
